Support GNMI native write with origin
